+++ content/includes/widgets/page-inspector-widget/spec/spec_runner.js
@@ -1,0 +1,15 @@
+/**
+ * Copyright 2015
+ * Released under the MIT license
+ */
+( function( global ) {
+   'use strict';
+   global.laxarSpec = {
+      title: 'ax-page-inspector-widget Specification',
+      tests: [
+         'ax-page-inspector-widget.spec'
+      ],
+      testRunner: 'laxar-mocks',
+      jasmineMajorVersion: 2
+  };
+} )( this );